requirements.txt
fastapi 
uvicorn 
openapi
python-dotenv
requests
streamlit

